CLIM-592 moved pydantic models to data_models, linted tests and uv lock
